Split files into smaller files
